% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qpcr}
\alias{qpcr}
\title{qpcr}
\format{
a data.frame
\describe{
\item{Log}{value1}
\item{Cq}{value2}
}
}
\usage{
qpcr
}
\description{
qpcr
}
\section{Functions}{
\itemize{
\item \code{qpcr}: data for Five-parameter log-logistic models testing
}}

\examples{
data(qpcr)
head(qpcr)
}
\keyword{datasets}
