% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four-parameter_loglogistic_model.R
\name{fourPL}
\alias{fourPL}
\title{Fitting Four-parameter log-logistic models and Ploting}
\usage{
fourPL(
  stan,
  res,
  stan.x = NULL,
  stan.y = NULL,
  toppt = FALSE,
  ppt.fw = 5,
  ppt.fh = 4,
  predict = FALSE,
  predict.inverse = TRUE
)
}
\arguments{
\item{stan}{dataFrame for fitting four-parameter log-logistic models}

\item{res}{numeric vector used for prediction}

\item{stan.x}{the name used for x axis of the fitting plot}

\item{stan.y}{the name used for y axis of the fitting plot}

\item{toppt}{TRUE or FALSE}

\item{ppt.fw}{figure width in ppt}

\item{ppt.fh}{figure heigh in ppt}

\item{predict}{TRUE or FALSE}

\item{predict.inverse}{TRUE or FALSE}
}
\value{
a list containing fit and plot
}
\description{
Fitting four-parameter log-logistic models.
Plotting the curve and export to ppt
The (inverse) prediction for the new data is provided.
}
\author{
Pan Gao
}
