% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear}
\alias{linear}
\title{Fitting Linear Models and Ploting}
\usage{
linear(
  stan,
  res,
  stan.x = NULL,
  stan.y = NULL,
  ppt.fw = 5,
  ppt.fh = 4,
  toppt = FALSE,
  predict = TRUE,
  predict.inverse = FALSE
)
}
\arguments{
\item{stan}{dataFrame for fitting linear polynomial regression}

\item{res}{numeric vector used for prediction}

\item{stan.x}{the name used for x axis of the fitting plot}

\item{stan.y}{the name used for y axis of the fitting plot}

\item{ppt.fw}{figure width in ppt}

\item{ppt.fh}{figure heigh in ppt}

\item{toppt}{TRUE or FALSE}

\item{predict}{TRUE or FALSE}

\item{predict.inverse}{TRUE or FALSE}
}
\value{
a list containing fit and plot
}
\description{
Fitting data with linear polynomial regression.
Plotting fitting curve and export to ppt.
The (inverse) prediction for new data
}
\author{
Pan Gao
}
